Tests commit, move fixtures & clean imports
