+++ tests/deepprofiler/dataset/test_utils.py
@@ -1,5 +1,4 @@
 import deepprofiler.dataset.utils
-import pytest
 import sys
 import os
 from io import StringIO
