+++ tests/plugins/crop_generators/test_crop_generator.py
@@ -1,5 +1,3 @@
-import pytest
-
 import deepprofiler.imaging.cropping
 import plugins.crop_generators.crop_generator
